+++ src/ic-cdk/src/api/stable/mod.rs
@@ -128,7 +128,7 @@ pub fn stable_bytes() -> Vec<u8> {
     // `vec`, being mutable and allocated to `size` bytes, is safe to pass to ic0.stable_read with no offset.
     // ic0.stable_read writes to all of `vec[0..size]`, so `set_len` is safe to call with the new size.
     unsafe {
-        ic0::stable_read(vec.as_ptr() as i32, 0, size as i32);
+        ic0::stable_read(vec.as_ptr() as usize, 0, size as u32);
         vec.set_len(size);
     }
     vec
